GFX macros change

+++ Ironware/DxgiInfoManager.h
@@ -22,6 +22,28 @@
 #include <string>
 #include <wrl.h>
 
+/*!
+ * \class DxgiInfoManager
+ *
+ * \ingroup DEV
+ *
+ * \brief A helper class that is used for getting DXGI error information
+ *
+ * TODO:
+ *
+ * \note Call Set every time you want to capture an error. Call Set => DirectXFunc => GetMessages
+ * * Set moves the message index to the present so that it will show only the messages related to the
+ * * function call
+ *
+ * \author Yernar Aldabergenov
+ *
+ * \version 1.0
+ *
+ * \date September 2020
+ *
+ * Contact: yernar.aa@gmail.com
+ *
+ */
 class DxgiInfoManager
 {
 public:
@@ -34,6 +56,6 @@ public:
 	std::vector<std::wstring> GetMessages() const;
 
 private:
-	unsigned long long next = 0u;
+	uint64_t next = 0u;
 	Microsoft::WRL::ComPtr<IDXGIInfoQueue> pDxgiInfoQueue;
 };
